move instrumentation related classes into own sub-package
