+++ src/main/groovy/com/jvoegele/gradle/tasks/android/instrumentation/InstrumentationTestsTask.groovy
@@ -1,6 +1,8 @@
-package com.jvoegele.gradle.tasks.android
+package com.jvoegele.gradle.tasks.android.instrumentation
+
 
 import com.jvoegele.gradle.plugins.android.AndroidPlugin;
+import com.jvoegele.gradle.tasks.android.AdbExec;
 import com.jvoegele.gradle.tasks.android.exceptions.InstrumentationTestsFailedException;
 import com.jvoegele.gradle.tasks.android.exceptions.AdbErrorException;
